Revert "i2c: efi: avoid 64-bit division"
Tomas Marek explained why 64-bit division was used in the first place:

> The issue here is that max_len, which represents the MaximumReceiveBytes
> or MaximumTransmitBytes in EFI I2C capabilities, is 32-bit. In some EFI
> implementations (including the one I use), the EFI I2C MaximumReceiveBytes
> and MaximumTransmitBytes are set to the maximum value (0xFFFFFFFF). This
> causes msg->len + max_len - 1 to overflow, leading to an incorrect number
> of returned operations. To address this, the msg->len is extended to 64 bits.

see https://lore.barebox.org/barebox/Zr2XxoD4Gce7bTB4@debian/

This reverts commit 4e01eb4889ffcebdf86051e8ee4eecfbfb17129c.

+++ drivers/i2c/busses/i2c-efi.c
@@ -114,7 +114,7 @@ static unsigned int efi_i2c_msg_op_cnt(const struct efi_i2c_priv *i2c_priv,
 
 	max_len = efi_i2c_max_len(i2c_priv, msg);
 
-	return (msg->len + max_len - 1) / max_len;
+	return ((u64)msg->len + max_len - 1) / max_len;
 }
 
 static unsigned int efi_i2c_req_op_cnt(
